Renamed TypeSerializer to PropertySerializer
