+++ src/PropeertySerializer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace EventSauce\ObjectHydrator;
 
-interface TypeSerializer
+interface PropeertySerializer
 {
     public function serialize(mixed $value, ObjectSerializer $serializer): mixed;
 }
